Committing two R scripts
Committing two R scripts at the same time: some changes to the initial
test script, and added a dummy 'readdata.R' script (which actually
doesn't do anything - just a matter of having a couple of scripts, to
see a list instead of single files). Also typing a bit more text in the
commit message to see how this text behaves; there seems to be a hard
return (after 80 characters?) combined with a wrap at the end of the
line in the window.